+++ sched/semaphore/semaphore.h
@@ -31,7 +31,6 @@
 #include <nuttx/compiler.h>
 #include <nuttx/semaphore.h>
 #include <nuttx/sched.h>
-#include <nuttx/atomic.h>
 
 #include <stdint.h>
 #include <stdbool.h>
@@ -39,8 +38,6 @@
 /****************************************************************************
  * Pre-processor Definitions
  ****************************************************************************/
-
-#define NXSEM_COUNT(s) ((FAR atomic_t *)&(s)->semcount)
 
 /****************************************************************************
  * Public Function Prototypes
