+++ syscall/syscall.csv
@@ -92,13 +92,13 @@
 "nxsem_destroy","nuttx/semaphore.h","","int","FAR sem_t *"
 "nxsem_getprioceiling","nuttx/semaphore.h","defined(CONFIG_PRIORITY_PROTECT)","int","FAR const sem_t *","FAR int *"
 "nxsem_open","nuttx/semaphore.h","defined(CONFIG_FS_NAMED_SEMAPHORES)","int","FAR sem_t **","FAR const char *","int","...","mode_t","unsigned int"
-"nxsem_post","nuttx/semaphore.h","","int","FAR sem_t *"
+"nxsem_post_slow","nuttx/semaphore.h","","int","FAR sem_t *"
 "nxsem_set_protocol","nuttx/semaphore.h","defined(CONFIG_PRIORITY_INHERITANCE)","int","FAR sem_t *","int"
 "nxsem_setprioceiling","nuttx/semaphore.h","defined(CONFIG_PRIORITY_PROTECT)","int","FAR sem_t *","int","FAR int *"
 "nxsem_timedwait","nuttx/semaphore.h","","int","FAR sem_t *","FAR const struct timespec *"
-"nxsem_trywait","nuttx/semaphore.h","","int","FAR sem_t *"
+"nxsem_trywait_slow","nuttx/semaphore.h","","int","FAR sem_t *"
 "nxsem_unlink","nuttx/semaphore.h","defined(CONFIG_FS_NAMED_SEMAPHORES)","int","FAR const char *"
-"nxsem_wait","nuttx/semaphore.h","","int","FAR sem_t *"
+"nxsem_wait_slow","nuttx/semaphore.h","","int","FAR sem_t *"
 "open","fcntl.h","","int","FAR const char *","int","...","mode_t"
 "pgalloc", "nuttx/arch.h", "defined(CONFIG_BUILD_KERNEL)", "uintptr_t", "uintptr_t", "unsigned int"
 "pipe2","unistd.h","defined(CONFIG_PIPES) && CONFIG_DEV_PIPE_SIZE > 0","int","int [2]|FAR int *","int"
